ca compile

+++ box.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   box.c                                              :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: alhote <alhote@student.42.fr>              +#+  +:+       +#+        */
+/*   By: snicolet <snicolet@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/05/23 21:03:45 by snicolet          #+#    #+#             */
-/*   Updated: 2016/05/31 22:04:59 by alhote           ###   ########.fr       */
+/*   Updated: 2016/06/01 13:59:36 by snicolet         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,12 +18,12 @@ int		raybox_check(t_ray *r, t_box *box)
 	double	tmax;
 	double	temp[2];
 
-	tmin = -INFINITY;
-	tmax = INFINITY;
+	tmin = (double)-INFINITY;
+	tmax = (double)INFINITY;
 	if (r->dir.x != 0.0f)
 	{
-		temp[0] = (box->xmin - r->start.x) / r->dir.x;
-		temp[1] = (box->xmax - r->start.x) / r->dir.x;
+		temp[0] = (double)((box->xmin - r->start.x) / r->dir.x);
+		temp[1] = (double)((box->xmax - r->start.x) / r->dir.x);
 		tmin = fmax(tmin, fmin(temp[0], temp[1]));
 		tmax = fmin(tmax, fmax(temp[0], temp[1]));
 		if (tmin > tmax)
@@ -31,8 +31,8 @@ int		raybox_check(t_ray *r, t_box *box)
 	}
 	if (r->dir.y != 0.0f)
 	{
-		temp[0] = (box->ymin - r->start.y) / r->dir.y;
-		temp[1] = (box->ymax - r->start.y) / r->dir.y;
+		temp[0] = (double)((box->ymin - r->start.y) / r->dir.y);
+		temp[1] = (double)((box->ymax - r->start.y) / r->dir.y);
 		tmin = fmax(tmin, fmin(temp[0], temp[1]));
 		tmax = fmin(tmax, fmax(temp[0], temp[1]));
 		if (tmin > tmax)
@@ -40,8 +40,8 @@ int		raybox_check(t_ray *r, t_box *box)
 	}
 	if (r->dir.z != 0.0f)
 	{
-		temp[0] = (box->zmin - r->start.z) / r->dir.z;
-		temp[1] = (box->zmax - r->start.z) / r->dir.z;
+		temp[0] = (double)((box->zmin - r->start.z) / r->dir.z);
+		temp[1] = (double)((box->zmax - r->start.z) / r->dir.z);
 		tmin = fmax(tmin, fmin(temp[0], temp[1]));
 		tmax = fmin(tmax, fmax(temp[0], temp[1]));
 		if (tmin > tmax)
